(fix) name checking + decl & user func overwriting
The name checking now throws an error if the object name is any of the reserved keywords (forgot to add that when changing to regex version of this function).
You could do this: "decl auto int t 5" and it would work. Checking
number of words now.
You can enable user-function-overwriting with he setting "user function overwriting".